requirements.txt
streamlit==1.28.0
transformers>=4.40.0
torch
torchvision==0.15.2
pillow==10.0.0
huggingface-hub==0.19.4
accelerate==0.25.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6